Implemented a nested do While loop inside a while loop to convert the username to lowercase so the username isn't case-sensitive. The loop now always prints out username until the user has entered their username.
I also fixed the problem where if a user got their password wrong on the first attempt and correct on the second attempt the program would ask for the username again and not grant access. To combat this I removed the if statement from the do while loop and used a boolean to set accessGranted to true if the username and password was correct which would end the loop.